19 October 2020, Just some changes in structure not in content. Made a new file for the linked list of component blocks (the organisational level that contains multiple tokens that have been grouped), as these will have their own linked list functions. Have removed most of the .h files and moved almost all the fsm related includes into the more general parse.h, after having tried to keep the tokeniser and the fsm handler separate initially. Have also grouped structs and prototypes within parse.h by 'section' of the program and added comments to keep them separate. This means a lot of changes in almost every file, as the includes of each of these files has been changed.
