+++ includes/parse.h
@@ -6,14 +6,18 @@
 /*   By: jsaariko <jsaariko@student.codam.nl>         +#+                     */
 /*                                                   +#+                      */
 /*   Created: 2020/10/14 14:04:54 by jsaariko      #+#    #+#                 */
-/*   Updated: 2020/10/19 16:07:48 by limartin      ########   odam.nl         */
+/*   Updated: 2020/10/19 18:51:32 by limartin      ########   odam.nl         */
 /*                                                                            */
 /* ************************************************************************** */
 
 #ifndef PARSE_H
 # define PARSE_H
 
-# include <stdlib.h>//
+# include <stdlib.h>
+
+/*
+** Tokeniser structs & functions
+*/
 
 typedef struct	s_token
 {
@@ -27,5 +31,50 @@ t_token	*create_token(char *tokens, int j, size_t len);
 void	free_matrix(char **tokens);
 void	parse_input(const char *input);
 void	manage_fsm(t_token *tokens);
+
+/*
+** Composition-block structs & functions
+*/
+
+typedef enum	s_composition_type
+{
+	type_separator,
+	type_command,
+	type_option,
+	type_arg
+}				t_composition_type;
+
+typedef struct	s_icomp
+{
+	char			*sep;
+	char			*cmd;
+	char			*opt;
+	char			*arg;
+	struct s_icomp	*next;
+}				t_icomp;
+
+int					ft_compconst(t_icomp *tonull);
+
+/*
+** FSM structs & functions
+*/
+
+typedef enum	s_transition_code
+{
+	state_padding,
+	state_command,
+	state_error
+}				t_transition_code;
+
+t_transition_code	sh_entry_state(t_token *this, t_icomp *icur);
+t_transition_code	sh_command_state(t_token *this, t_icomp *icur);
+t_transition_code	sh_error_state(t_token *this, t_icomp *icur);
+
+typedef struct	s_transition_obj
+{
+	t_transition_code	(*orig_state)(t_token *this, t_icomp *icur);
+	t_transition_code	transition_code;
+	t_transition_code	(*next_state)(t_token *this, t_icomp *icur);
+}				t_transition_obj;
 
 #endif
