+++ srcs/entry_state.c
@@ -6,13 +6,11 @@
 /*   By: limartin <limartin@student.codam.nl>         +#+                     */
 /*                                                   +#+                      */
 /*   Created: 2020/10/15 18:10:03 by limartin      #+#    #+#                 */
-/*   Updated: 2020/10/15 20:11:19 by limartin      ########   odam.nl         */
+/*   Updated: 2020/10/19 18:52:34 by limartin      ########   odam.nl         */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "minishell.h"
 #include "parse.h"
-#include "fsm.h"
 
 t_transition_code sh_entry_state(t_token *this, t_icomp *icur)
 {
