+++ srcs/tokenizer.c
@@ -6,7 +6,7 @@
 /*   By: jsaariko <jsaariko@student.codam.nl>         +#+                     */
 /*                                                   +#+                      */
 /*   Created: 2020/10/15 11:38:56 by jsaariko      #+#    #+#                 */
-/*   Updated: 2020/10/16 14:53:32 by jsaariko      ########   odam.nl         */
+/*   Updated: 2020/10/19 18:53:31 by limartin      ########   odam.nl         */
 /*                                                                            */
 /* ************************************************************************** */
